Extracted common shared code in abstract superclass
